+++ sec_certs/utils/helpers.py
